improve installation for other platforms
