+++ shell.nix
